+++ front/Window.java
@@ -6,11 +6,11 @@ import javax.swing.*;
 import java.awt.*;
 
 public class Window extends Canvas {
-        public Window(int width, int height, String title, Simulation simulation){
+        public Window(double width, double height, String title, Simulation simulation){
             JFrame frame = new JFrame(title);
-            frame.setPreferredSize(new Dimension(width, height));
-            frame.setMaximumSize(new Dimension(width, height));
-            frame.setMinimumSize(new Dimension(width, height));
+            frame.setPreferredSize(new Dimension((int)width, (int)height));
+            frame.setMaximumSize(new Dimension((int)width, (int)height));
+            frame.setMinimumSize(new Dimension((int)width, (int)height));
 
             frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
             frame.setResizable(false);
